+++ include/jdksavdecc_descriptor_storage_gen.h
@@ -128,7 +128,7 @@ uint16_t jdksavdecc_descriptor_storage_gen_read_descriptor(
         uint16_t configuration_number,
         uint16_t descriptor_type,
         uint16_t descriptor_index,
-        uint16_t *result_buffer,
+        uint8_t *result_buffer,
         uint16_t result_buffer_len);
 
 /// Write the descriptor data for the specified configuration, descriptor_type and
@@ -139,7 +139,7 @@ uint16_t jdksavdecc_descriptor_storage_gen_write_descriptor(
         uint16_t configuration_number,
         uint16_t descriptor_type,
         uint16_t descriptor_index,
-        uint16_t const *descriptor_data,
+        uint8_t const *descriptor_data,
         uint16_t descriptor_data_length);
 
 uint16_t jdksavdecc_descriptor_storage_gen_read_localized_string(
